requirements/dev.txt
-r base.txt
django-debug-toolbar==1.3.0
pyinotify==0.9.5
